PRU/requirements.txt
opencv-python
matplotlib
h5py
-e git+https://github.com/lmkoch/medpy/#egg=medpy
revtorch
scikit-image
nibabel
comet-ml